# jaxrs-spec-generated/swagger.json
{
  "openapi" : "3.0.7",
  "info" : {
    "title" : "Metamind API",
    "description" : "Brain spec for Metamind.",
    "version" : "2.0.0"
  },
  "servers" : [ {
    "url" : "/v2"
  } ],
  "security" : [ {
    "BearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/trainingMaterials" : {
      "get" : {
        "tags" : [ "TrainingMaterials" ],
        "summary" : "List trainingMaterials",
        "description" : "List training materials",
        "operationId" : "listTrainingMaterials",
        "parameters" : [ {
          "name" : "storyId",
          "in" : "query",
          "description" : "story id",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "type",
          "in" : "query",
          "description" : "training material type",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "$ref" : "#/components/schemas/TrainingMaterialType"
          }
        }, {
          "name" : "visibility",
          "in" : "query",
          "description" : "training material visibility",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "$ref" : "#/components/schemas/TrainingMaterialVisibility"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "TrainingMaterials",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TrainingMaterial"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "TrainingMaterials" ],
        "summary" : "Create a trainingMaterial",
        "description" : "Create training material",
        "operationId" : "createTrainingMaterial",
        "requestBody" : {
          "description" : "Payload",
          "content" : {
            "application/json;charset=utf-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/TrainingMaterial"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Created trainingMaterial",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrainingMaterial"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/trainingMaterials/{trainingMaterialId}" : {
      "get" : {
        "tags" : [ "TrainingMaterials" ],
        "summary" : "Finds a trainingMaterial",
        "description" : "Finds a training material",
        "operationId" : "findTrainingMaterial",
        "parameters" : [ {
          "name" : "trainingMaterialId",
          "in" : "path",
          "description" : "trainingMaterial id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Found trainingMaterial",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrainingMaterial"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "TrainingMaterials" ],
        "summary" : "Updates trainingMaterial",
        "description" : "Updates a training material",
        "operationId" : "updateTrainingMaterial",
        "parameters" : [ {
          "name" : "trainingMaterialId",
          "in" : "path",
          "description" : "trainingMaterial id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "description" : "Payload",
          "content" : {
            "application/json;charset=utf-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/TrainingMaterial"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Updated trainingMaterial",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrainingMaterial"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "TrainingMaterials" ],
        "summary" : "Delete trainingMaterial",
        "description" : "Delete a training material",
        "operationId" : "deleteTrainingMaterial",
        "parameters" : [ {
          "name" : "trainingMaterialId",
          "in" : "path",
          "description" : "trainingMaterial id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful removal"
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TrainingMaterial" : {
        "required" : [ "name", "text" ],
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "intent id.",
            "format" : "uuid",
            "readOnly" : true
          },
          "type" : {
            "$ref" : "#/components/schemas/TrainingMaterialType"
          },
          "storyId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "text" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "visibility" : {
            "$ref" : "#/components/schemas/TrainingMaterialVisibility"
          },
          "createdAt" : {
            "type" : "string",
            "description" : "Creation time",
            "format" : "date-time",
            "readOnly" : true
          },
          "modifiedAt" : {
            "type" : "string",
            "description" : "Last modification time",
            "format" : "date-time",
            "readOnly" : true
          }
        }
      },
      "TrainingMaterialType" : {
        "type" : "string",
        "enum" : [ "INTENTREGEX", "INTENTOPENNLPDOCCAT", "VARIABLEOPENNLPNER", "VARIABLEOPENNLPREGEX" ]
      },
      "TrainingMaterialVisibility" : {
        "type" : "string",
        "enum" : [ "STORY", "LOCAL" ]
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "The error message",
            "readOnly" : true
          }
        }
      },
...
    },
    "securitySchemes" : {
      "BearerAuth" : {
        "type" : "http",
        "scheme" : "bearer"
      }
    }
  }
}